:recycle: Renamed card-view to task-card-view
