+++ src/app/_shared/tasks/card-view/task-card-view.component.css
