+++ src/app/app-routing.module.ts
@@ -8,7 +8,7 @@ import {GraphComponent} from './_shared/graph/graph.component';
 
 import {TaskListViewComponent} from './_shared';
 import {GraphViewComponent} from './_shared/tasks/graph-view/graph-view.component';
-import {TaskCardViewComponent} from './_shared/tasks/card-view/card-view.component';
+import {TaskCardViewComponent} from './_shared/tasks/card-view/task-card-view.component';
 import {ServicesCardViewComponent} from './_shared/services/card-view/services-card-view.component';
